move supporting packages under pkg
pkg is a go convention for repos that are programs i.e. sbcli, broker,
etc. When writing a go library, you put the files at the top level like
bundle-lib.
